pulled Lecture into own class; small refactoring of LectureGridPagingAdapter
